Salting team id

+++ Exporter/Backend/src/modules/transport_layer/domain_value/arena_team.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone, Serialize)]
 pub struct ArenaTeam {
-  pub team_id: u32,
+  pub team_id: u64,
   pub name: String,
   pub team_type: u8,
   pub team_rating: u16,
